+++ BingeBuddyNg/BingeBuddyNg.Services/User/Commands/SetFriendMuteStateCommand.cs
@@ -1,7 +1,7 @@
 ﻿using MediatR;
 using System;
-using System.Collections.Generic;
-using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace BingeBuddyNg.Services.User.Commands
 {
@@ -17,5 +17,29 @@ namespace BingeBuddyNg.Services.User.Commands
         public string UserId { get; }
         public string FriendUserId { get; }
         public bool MuteState { get; }
+    }
+
+    public class SetFriendMuteStateCommandHandler : IRequestHandler<SetFriendMuteStateCommand>
+    {
+        private readonly IUserRepository userRepository;
+
+        public SetFriendMuteStateCommandHandler(IUserRepository userRepository)
+        {
+            this.userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
+        }
+
+
+        public async Task<Unit> Handle(SetFriendMuteStateCommand request, CancellationToken cancellationToken)
+        {
+            var user = await userRepository.FindUserAsync(request.UserId);
+            user.SetFriendMuteState(request.FriendUserId, request.MuteState);
+
+            var mutedUser = await userRepository.FindUserAsync(request.FriendUserId);
+            mutedUser.SetMutedByFriendState(request.UserId, request.MuteState);
+
+            Task.WaitAll(userRepository.UpdateUserAsync(user), userRepository.UpdateUserAsync(mutedUser));
+
+            return Unit.Value;
+        }
     }
 }
