+++ BingeBuddyNg/BingeBuddyNg.Services/User/Commands/UpdateUserProfileImageCommand.cs
@@ -1,8 +1,11 @@
-﻿using MediatR;
+﻿using BingeBuddyNg.Services.Activity;
+using BingeBuddyNg.Services.Infrastructure;
+using MediatR;
 using Microsoft.AspNetCore.Http;
 using System;
-using System.Collections.Generic;
-using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using static BingeBuddyNg.Shared.Constants;
 
 namespace BingeBuddyNg.Services.User.Commands
 {
@@ -16,5 +19,31 @@ namespace BingeBuddyNg.Services.User.Commands
 
         public string UserId { get; }
         public IFormFile Image { get; }
+    }
+
+    public class UpdateUserProfileImageCommandHandler : IRequestHandler<UpdateUserProfileImageCommand>
+    {
+        private readonly IUserRepository userRepository;
+        private readonly IActivityRepository activityRepository;
+        private readonly IStorageAccessService storageAccessService;
+
+        public UpdateUserProfileImageCommandHandler(IUserRepository userRepository, IActivityRepository activityRepository, IStorageAccessService storageAccessService)
+        {
+            this.userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
+            this.activityRepository = activityRepository ?? throw new ArgumentNullException(nameof(activityRepository));
+            this.storageAccessService = storageAccessService ?? throw new ArgumentNullException(nameof(storageAccessService));
+        }
+
+        public async Task<Unit> Handle(UpdateUserProfileImageCommand request, CancellationToken cancellationToken)
+        {
+            var user = await this.userRepository.FindUserAsync(request.UserId);
+            using (var stream = request.Image.OpenReadStream())
+            {
+                await storageAccessService.SaveFileInBlobStorage(ContainerNames.ProfileImages, request.UserId, stream);
+                var activity = Activity.Activity.CreateProfileImageUpdateActivity(request.UserId, user.Name);
+                await activityRepository.AddActivityAsync(activity);
+            }
+            return Unit.Value;
+        }
     }
 }
