+++ BingeBuddyNg/BingeBuddyNg.Services/User/Querys/GetAllUsersQuery.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace BingeBuddyNg.Services.User.Querys
 {
@@ -13,5 +15,31 @@ namespace BingeBuddyNg.Services.User.Querys
         }
 
         public string FilterText { get; }
+    }
+
+    public class GetAllUsersQueryHandler : IRequestHandler<GetAllUsersQuery, List<UserInfoDTO>>
+    {
+        private readonly IUserRepository userRepository;
+
+        public GetAllUsersQueryHandler(IUserRepository userRepository)
+        {
+            this.userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
+        }
+
+        public async Task<List<UserInfoDTO>> Handle(GetAllUsersQuery request, CancellationToken cancellationToken)
+        {
+            var users = await this.userRepository.GetUsersAsync();
+
+            var userInfo = users.Select(u => new UserInfoDTO() { UserId = u.Id, UserName = u.Name }).ToList();
+
+            // TODO: Should soon be improved!
+            if (!string.IsNullOrEmpty(request.FilterText))
+            {
+                string lowerFilter = request.FilterText.ToLower();
+                userInfo = userInfo.Where(u => u.UserName.ToLower().Contains(lowerFilter)).ToList();
+            }
+
+            return userInfo;
+        }
     }
 }
