+++ BingeBuddyNg/BingeBuddyNg.Services/User/Querys/GetUserQuery.cs
@@ -1,7 +1,7 @@
 ﻿using MediatR;
 using System;
-using System.Collections.Generic;
-using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace BingeBuddyNg.Services.User.Querys
 {
@@ -13,5 +13,39 @@ namespace BingeBuddyNg.Services.User.Querys
         }
 
         public string UserId { get; }
+    }
+
+    public class GetUserQueryHandler : IRequestHandler<GetUserQuery, UserDTO>
+    {
+        private readonly IUserRepository userRepository;
+
+        public GetUserQueryHandler(IUserRepository userRepository)
+        {
+            this.userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
+        }
+
+        public async Task<UserDTO> Handle(GetUserQuery request, CancellationToken cancellationToken)
+        {
+            var user = await this.userRepository.FindUserAsync(request.UserId);
+            if (user == null)
+            {
+                return null;
+            }
+
+            return new UserDTO()
+            {
+                Id = user.Id,
+                Name = user.Name,
+                PushInfo = user.PushInfo,
+                ProfileImageUrl = user.ProfileImageUrl,
+                Gender = user.Gender,
+                Language = user.Language,
+                LastOnline = user.LastOnline,
+                Weight = user.Weight,
+                CurrentVenue = user.CurrentVenue,
+                Friends = user.Friends,
+                MutedFriendUserIds = user.MutedFriendUserIds
+            };
+        }
     }
 }
